Finished modularization. Got the wwwSocket.test.js working as it needed the TSL .pems defined to work correctly with the wss socket server. I will have to use a similar approach in future tests
