+++ routes/radio.js
@@ -6,6 +6,7 @@ import icy from 'icy'; // Import the icy module
 import { wss } from '../bin/www.js';
 import { fetchSpotifyTrackMetadata } from '../services/spotifyTrackSearch.js';
 import { broadcastMetadata } from '../services/utilities/webSocketUtils.js';
+import { monitorMetadata } from '../services/radio/monitorMetadata.js';
 
 const router = express.Router();
 
@@ -19,69 +20,6 @@ const axiosInstance = axios.create({
     }),
 });
 
-/**
- * Monitor metadata and broadcast to WebSocket clients.
- */
-const monitorMetadata = (streamUrl) => {
-  icy.get(streamUrl, (res) => {
-    console.log(`Connected to stream: ${streamUrl}`);
-
-    let currentTrack = null;
-    let isConnectionAlive = true;
-
-    setInterval(() => {
-      if (!isConnectionAlive) {
-        console.warn('Stream connection lost. Reconnecting...');
-        res.destroy();
-        monitorMetadata(streamUrl);
-      }
-      isConnectionAlive = false;
-    }, 30000);
-
-    res.on('data', (chunk) => {
-      isConnectionAlive = true; // Connection is alive if data is received
-    });
-
-    res.on('metadata', (metadata) => {
-      console.log('Metadata event triggered');
-      try {
-        const parsedMetadata = icy.parse(metadata);
-        
-
-        const trackInfo = parsedMetadata?.StreamTitle || 'Unknown';
-        if (trackInfo !== currentTrack) {
-          currentTrack = trackInfo;
-          const metadataObj = extractMetadata(trackInfo);
-
-          broadcastMetadata(metadataObj);
-
-          enrichMetadata(metadataObj).then((enrichedMetadata) => {
-            broadcastMetadata(enrichedMetadata);
-          }).catch((error) => {
-            console.error('Error enriching metadata:', error.message);
-          });
-        } else {
-          console.log('Duplicate metadata detected, skipping broadcast');
-        }
-      } catch (error) {
-        console.error('Error processing metadata:', error.message);
-      }
-    });
-
-    res.on('end', () => {
-      console.log('Stream ended. Reconnecting...');
-      setTimeout(() => monitorMetadata(streamUrl), 5000);
-    });
-
-    res.on('error', (err) => {
-      console.error('Stream error:', err.message);
-      setTimeout(() => monitorMetadata(streamUrl), 5000);
-    });
-  }).on('error', (err) => {
-    console.error('Failed to connect to stream:', err.message);
-    setTimeout(() => monitorMetadata(streamUrl), 5000);
-  });
-};
 
 
 const log = (action, data) => {
@@ -295,5 +233,5 @@ router.get('/api/metadata', async (req, res) => {
     }
   });
 
-export {extractMetadata, broadcastMetadata, enrichMetadata, monitorMetadata}
+export {extractMetadata, enrichMetadata}
 export { router as radioRouter };
