+++ services/utilities/webSocketUtils.js
@@ -1,6 +1,3 @@
-// websocketUtils.js
-import { WebSocket } from 'ws';
-
 let wssReference = null;
 
 export const setWebSocketServer = (wss) => {
@@ -8,20 +5,27 @@ export const setWebSocketServer = (wss) => {
 };
 
 export const getWebSocketServer = () => {
+  if (!wssReference) {
+    throw new Error('WebSocket Server is not initialized');
+  }
   return wssReference;
 };
 
 export const broadcastMetadata = (metadata) => {
-  const wss = getWebSocketServer();
-  if (!wss || !metadata.currentTrack || metadata.currentTrack === 'Unknown') {
-    console.warn('No valid metadata to broadcast or WebSocket not initialized');
+  if (!wssReference) {
+    console.warn('WebSocket server not initialized; skipping broadcast');
+    return;
+  }
+
+  if (!metadata.currentTrack || metadata.currentTrack === 'Unknown') {
+    console.warn('No valid metadata to broadcast');
     return;
   }
 
   metadata.startTime = metadata.startTime || Date.now();
 
-  wss.clients.forEach((client) => {
-    if (client.readyState === WebSocket.OPEN) {
+  wssReference.clients.forEach((client) => {
+    if (client.readyState === 1) { // WebSocket.OPEN === 1
       try {
         client.send(JSON.stringify(metadata));
       } catch (error) {
